Add method sending early partial request to devtools and move adding response logic to send_response_to_devtools
